requirements.txt
# AI Wallpaper Generator Requirements

# Core dependencies
click>=8.0.0
pyyaml>=6.0
requests>=2.25.0
Pillow>=9.0.0
psutil>=5.8.0

# PyTorch and related (for FLUX and SDXL)
torch>=2.0.0
torchvision>=0.15.0
torchaudio

# Diffusers and transformers for AI models
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.25.0
safetensors>=0.4.0
omegaconf>=2.3.0
einops>=0.7.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# For Real-ESRGAN (install separately)
# basicsr
# facexlib
# gfpgan

# OpenAI for DALL-E and GPT models
openai>=1.0.0

# Optional but recommended
xformers>=0.0.23  # For memory efficiency
bitsandbytes>=0.41.0  # For 8-bit optimizations

# Development tools (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0